+++ pkg/nanolib/delay/src/main.ts
@@ -1,5 +1,7 @@
-export * from './queue-microtask.js';
-export * from './request-animation-frame.js';
-export * from './request-idle-callback.js';
-export * from './queue-macrotask.js';
+export * from './queue_microtask.js';
+export * from './request_animation_frame.js';
+export * from './request_idle_callback.js';
+export * from './queue_macrotask.js';
+export * from './request_next_render.js';
+export * from './schedule_idle_batch.js';
 export * from './delay.js';
